feat: Remove `commitlint` from the extension recommendations
Due to bad performance, we decided to delete
this extension from the recommendations
by the moment. If the extension fix their bugs
and increase its performance in the future
will be considered again.

Despite everything, we want to thank the
developer for his effort in this great idea,
let's hope he continues with it!