feat: Remove currency translations from the user database (#171)
Closes #166.

This PR remove the need of another user table for the currency names, saving space in the DB and improving its performance. Also, allows new languages without the need of other migrations. 

This new implementation has the drawback right now that if the user changes the language of the application later, the names of the currencies will not be translated (same behavior as with the categories).